+++ lib/promutils/duration.go
@@ -44,12 +44,3 @@ func (pd *Duration) Duration() time.Duration {
 	}
 	return pd.D
 }
-
-// ParseDuration parses duration string in Prometheus format
-func ParseDuration(s string) (time.Duration, error) {
-	ms, err := metricsql.DurationValue(s, 0)
-	if err != nil {
-		return 0, err
-	}
-	return time.Duration(ms) * time.Millisecond, nil
-}
