+++ lib/timeutil/duration.go
@@ -1,0 +1,16 @@
+package timeutil
+
+import (
+	"time"
+
+	"github.com/VictoriaMetrics/metricsql"
+)
+
+// ParseDuration parses duration string in Prometheus format
+func ParseDuration(s string) (time.Duration, error) {
+	ms, err := metricsql.DurationValue(s, 0)
+	if err != nil {
+		return 0, err
+	}
+	return time.Duration(ms) * time.Millisecond, nil
+}
